Add files via upload
At CodeAlpha, we're beginning with the fundamentals, starting with the Fibonacci sequence. The Fibonacci series is a sequence where each number is the sum of the two preceding numbers, defined by a mathematical recurrence relationship. In task 1 as part of the Python programming internship at CodeAlpha, i'll be tasked with developing a program capable of generating and manipulating Fibonacci sequences.